create resource package (#503)
