--- docs/source/api_reference/framework/resource.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium.framework.resource
