+++ docs/source/api_reference/framework/resource/index.rst
@@ -1,0 +1,11 @@
+===================
+Resource Management
+===================
+
+.. automodule:: vivarium.framework.resource
+
+.. toctree::
+   :maxdepth: 1
+   :glob:
+
+   *
