+++ docs/source/api_reference/framework/resource/manager.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium.framework.resource.manager
