+++ src/vivarium/framework/resource/exceptions.py
@@ -1,0 +1,7 @@
+from vivarium.exceptions import VivariumError
+
+
+class ResourceError(VivariumError):
+    """Error raised when a dependency requirement is violated."""
+
+    pass
